+++ packages/cli/__tests__/zosjobs/__integration__/modify/__scripts__/job/help.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+zowe zos-jobs modify job -h
+if [ $? -gt 0 ]
+then
+    exit $?
+fi
+
+zowe zos-jobs modify job --help --response-format-json
+exit $?