+++ packages/cli/__tests__/zosjobs/__integration__/modify/__scripts__/job/missing_jobid.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+zowe zos-jobs modify job
+exit $?