+++ packages/cli/__tests__/zosjobs/__system__/modify/__scripts__/job/bogus_jobid.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+zowe zos-jobs modify job JOB00000 --jobclass A
+RC=$?
+
+exit $?